Adapting recovery & refresh tests
